educational purpose/twitter scarping bot/requirements.txt
XlsxWriter==3.0.1
webdriver_manager==3.4.2
selenium==3.141.0
python_dateutil==2.8.2
